editing welcome.py, added menu.py for auto menu-generation
Script oldwelcome.py was the original welcome.py, which I was editing to
include wait_for_message timeouts. I realized there got to be too many temporary queues and dicts in this
implementation and decided to refactor. This script will not run as-is,
but it is near completion.

Current welcome.py is a new script refactoring of oldwelcome.py to use
database entries instead of temporary data structures (there got to be
too many temporary queues and dicts in the previous implementation).